backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
pandas
scikit-learn
joblib
numpy
python-dotenv
